Add comments

+++ src/interfaces/BankInterface.java
@@ -9,8 +9,8 @@ import java.util.Date;
 
 public interface BankInterface extends Remote {
     long login(String username, String password) throws RemoteException, InvalidLoginException;
-    double deposit(int accountnum, int amount, long sessionID) throws RemoteException, InvalidSessionException;
-    double withdraw(int accountnum, int amount, long sessionID) throws RemoteException, InvalidSessionException, InsufficientFundsException;
+    double deposit(int accountnum, double amount, long sessionID) throws RemoteException, InvalidSessionException;
+    double withdraw(int accountnum, double amount, long sessionID) throws RemoteException, InvalidSessionException, InsufficientFundsException;
     Account inquiry(int accountnum, long sessionID) throws RemoteException, InvalidSessionException;
     StatementInterface getStatement(int accountnum, Date from, Date to, long sessionID) throws RemoteException, InvalidSessionException, StatementException;
     Account accountDetails(long sessionID) throws RemoteException, InvalidSessionException;
